requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.6
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
itsdangerous==2.1.2
python-dotenv==0.21.1
WTForms==3.0.1
email-validator==1.3.1
SQLAlchemy==2.0.17
